test(role): add role tests to main testing file
